Move eachh APIs from page to a separated file
