+++ inginious/frontend/plugins/problem_bank/pages/api/search_task_api.py
@@ -1,0 +1,29 @@
+from .admin_api import AdminApi
+
+class SearchTaskApi(AdminApi):
+    def API_GET(self):
+        bank_course_ids = set(bank["courseid"]
+                              for bank in self.database.problem_banks.find())
+
+        tasks = []
+        for bank_course_id in bank_course_ids:
+            search_tasks = self.database.tasks_cache.aggregate([
+                {"$match":
+                     {
+                         "course_id": bank_course_id
+                     }
+                }
+            ])
+            for task in search_tasks:
+                dict = {"course_id": task["course_id"], "task_id": task["task_id"], "task_name": task["task_name"],
+                        "task_author": task["task_author"], "task_context": task["task_context"],
+                        "tags": self.parse_tags(task["tags"])}
+                tasks.append(dict)
+
+        return 200, sorted(tasks, key=lambda k: (k['course_id'], k['task_id']))
+
+    def parse_tags(self, tags):
+        parsed_tags = list()
+        for key, tag in tags.items():
+            parsed_tags.append(tag["name"])
+        return parsed_tags
